feat: Add stock creation endpoint
The code changes in this commit add a new endpoint to the `StockController` for creating a new stock. The endpoint accepts a `CreateStockRequestDto` object in the request body and maps it to a `Stock` model using the `ToStockFromCreateDTO` method in the `StockMappers` class. The created stock is then added to the database and the changes are saved.

This commit message follows the established convention of using a prefix to indicate the type of change (`feat` for feature) and provides a clear and concise description of the purpose of the code changes.